+++ src/pages/Observatories/Observatory.js
@@ -15,6 +15,7 @@ import ContributorCard from 'components/ContributorCard/ContributorCard';
 import PaperCard from 'components/PaperCard/PaperCard';
 import ComparisonCard from 'components/ComparisonCard/ComparisonCard';
 import EditObservatory from 'components/Observatory/EditObservatory';
+import RelatedResourcesCard from 'components/Observatory/RelatedResourcesCard';
 import AddResearchProblem from 'components/Observatory/AddResearchProblem';
 import NotFound from 'pages/NotFound';
 import PropTypes from 'prop-types';
@@ -27,6 +28,8 @@ import { getPaperData, getComparisonData } from 'utils';
 import { find } from 'lodash';
 import capitalize from 'capitalize';
 import { connect } from 'react-redux';
+import { filterObjectOfStatementsByPredicate } from 'utils';
+import { PREDICATES } from 'constants/graphSettings';
 
 class Observatory extends Component {
     constructor(props) {
@@ -36,7 +39,7 @@ class Observatory extends Component {
             error: null,
             label: '',
             description: '',
-            researchField: '',
+            researchField: null,
             isContributorsModalOpen: false,
             isLoading: false,
             isLoadingContributors: false,
@@ -128,11 +131,18 @@ class Observatory extends Component {
                 }).then(resourcesStatements => {
                     const comparisonsData = resourcesStatements.map(resourceStatements => {
                         const comparisonSubject = find(comparisons, { id: resourceStatements.id });
-                        return getComparisonData(
+                        const resources = filterObjectOfStatementsByPredicate(resourceStatements.statements, PREDICATES.RELATED_RESOURCES, false);
+                        const figures = filterObjectOfStatementsByPredicate(resourceStatements.statements, PREDICATES.RELATED_FIGURE, false);
+
+                        const data = getComparisonData(
                             resourceStatements.id,
                             resourceStatements && comparisonSubject.label ? comparisonSubject.label : 'No Title',
                             resourceStatements.statements
                         );
+
+                        data.resources = resources;
+                        data.figures = figures;
+                        return data;
                     });
                     this.setState({
                         comparisonsList: comparisonsData,
@@ -210,7 +220,7 @@ class Observatory extends Component {
 
                         <Container className="box rounded-lg clearfix pt-4 pb-4 pl-5 pr-5">
                             <h3>{this.state.label}</h3>
-                            {this.props.user && (
+                            {!!this.props.user && this.props.user.isCurationAllowed && (
                                 <Button
                                     color="darkblue"
                                     size="sm"
@@ -222,9 +232,14 @@ class Observatory extends Component {
                             )}
                             {this.state.description}
                             <br />
-                            <div className="flex-grow-1">
-                                <small>Research field: {this.state.researchField}</small>
-                            </div>
+                            {this.state.researchField && this.state.researchField.id && (
+                                <div className="flex-grow-1 mt-2">
+                                    Research field:
+                                    <Link className="ml-2" to={reverse(ROUTES.RESEARCH_FIELD, { researchFieldId: this.state.researchField.id })}>
+                                        {this.state.researchField && this.state.researchField.label}
+                                    </Link>
+                                </div>
+                            )}
                         </Container>
 
                         <Container>
@@ -232,7 +247,7 @@ class Observatory extends Component {
                                 <Col md={4} sm={12} style={{ minHeight: '300px' }} className="d-flex px-0 pr-3">
                                     <div className="box rounded-lg p-4 flex-grow-1">
                                         <h5>Research Problems</h5>
-                                        {this.props.user && (
+                                        {!!this.props.user && this.props.user.isCurationAllowed && (
                                             <Button
                                                 outline
                                                 size="sm"
@@ -408,6 +423,22 @@ class Observatory extends Component {
                         </Container>
 
                         <Container className="box rounded-lg p-4">
+                            <h5>Figures</h5>
+                            {!this.state.isLoadingComparisons ? (
+                                <div className="mb-4 mt-4">
+                                    {this.state.comparisonsList.length > 0 ? (
+                                        <RelatedResourcesCard figureStatements={this.state.comparisonsList} />
+                                    ) : (
+                                        <div className="text-center mt-4 mb-4">No Figures</div>
+                                    )}
+                                </div>
+                            ) : (
+                                <div className="text-center mt-4 mb-4">Loading figures ...</div>
+                            )}
+                        </Container>
+                        <br />
+
+                        <Container className="box rounded-lg p-4">
                             <h5>Comparisons</h5>
                             {!this.state.isLoadingComparisons ? (
                                 <div className="mb-4 mt-4">
@@ -485,7 +516,7 @@ Observatory.propTypes = {
             id: PropTypes.string.isRequired
         }).isRequired
     }).isRequired,
-    user: PropTypes.object
+    user: PropTypes.oneOfType([PropTypes.object, PropTypes.number])
 };
 
 export default connect(mapStateToProps)(Observatory);