+++ src/pages/Observatories/OrganizationObservatories.js
@@ -109,7 +109,7 @@ OrganizationObservatories.propTypes = {
             id: PropTypes.string.isRequired
         }).isRequired
     }).isRequired,
-    user: PropTypes.object
+    user: PropTypes.oneOfType([PropTypes.object, PropTypes.number])
 };
 
 export default connect(