+++ src/pages/Organizations/OrganizationDetails.js
@@ -159,7 +159,7 @@ class OrganizationDetails extends Component {
                                                 {this.state.label} {this.state.url && <Icon size="sm" icon={faExternalLinkAlt} />}
                                             </h4>
                                         </NavLink>
-                                        {this.props.user && this.props.user.id === this.state.createdBy && (
+                                        {!!this.props.user && (this.props.user.id === this.state.createdBy || this.props.user.isCurationAllowed) && (
                                             <div>
                                                 <Button
                                                     outline
@@ -287,7 +287,7 @@ OrganizationDetails.propTypes = {
             id: PropTypes.string.isRequired
         }).isRequired
     }).isRequired,
-    user: PropTypes.object
+    user: PropTypes.oneOfType([PropTypes.object, PropTypes.number])
 };
 
 const mapStateToProps = state => ({