+++ src/services/backend/observatories.js
@@ -40,6 +40,10 @@ export const getProblemsByObservatoryId = id => {
     return submitGetRequest(`${observatoriesUrl}${encodeURIComponent(id)}/problems`);
 };
 
+export const getObservatoriesByResearchFieldId = id => {
+    return submitGetRequest(`${observatoriesUrl}research-field/${encodeURIComponent(id)}/observatories`);
+};
+
 export const getObservatoriesStats = id => {
     return submitGetRequest(`${observatoriesUrl}stats/observatories`);
 };